Auth Service - Role repo

+++ auth-service/src/main/java/com/hrps/authservice/repository/RoleRepository.java
@@ -1,6 +1,7 @@
 package com.hrps.authservice.repository;
 
+import com.hrps.authservice.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
 }
